backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
pgvector==0.3.6
pydantic==2.10.4
python-dotenv==1.0.1
boto3==1.35.90
python-multipart==0.0.20
alembic==1.14.0
jinja2
reportlab
# ── 多轮对话模块新增依赖 ──────────────────────
redis[hiredis]>=5.0.0
sentence-transformers>=3.0.0
langdetect>=1.0.9
httpx>=0.27.0
# ── 번역 모듈 신규 의존성 ──────────────────────
librosa>=0.10.0
soundfile>=0.12.0
SpeechRecognition>=3.10.0
pydub>=0.25.1
